Add du repo Product pour gérer les id automatiquement et les mettre en cache si déjà renvoyé par l'API

+++ src/main/java/fr/isima/tp1/model/RuleRepository.java
@@ -6,8 +6,9 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 
+/* Repository pour les requêtes sur la table rule */
 @Repository
-public interface RuleRepository extends JpaRepository<Rule, Integer> {
+public interface RuleRepository extends JpaRepository<Rule, Long> {
     @Query("SELECT MAX(r.points) FROM Rule r WHERE  ?1 >= r.min_bound AND r.name = ?2")
     Integer findByNameAndValue(double value, String name);
 }
